requirements.txt
torch==2.3.1
flask==3.0.3
opencv-python ==4.10.0.82
pillow==10.3.0
torchvision==0.18.1
tensorflow==2.16.1
gdown==5.2.0
requests==2.32.3